requirements.txt
django
djangorestframework
rzpipe
angr
capstone -> pip install --pre --no-binary capstone capstone